+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -9,9 +9,9 @@ import org.apache.coyote.Processor;
 import org.apache.coyote.http11.controller.Controller;
 import org.apache.coyote.http11.dispatcher.RequestMapping;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.RequestFactory;
+import org.apache.coyote.http11.file.RequestFactory;
 import org.apache.coyote.http11.response.HttpResponse;
-import org.apache.coyote.http11.response.ResponseFactory;
+import org.apache.coyote.http11.file.ResponseFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
