+++ tomcat/src/main/java/org/apache/coyote/http11/file/MimeTypeMaker.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.file;
 
 import java.util.Arrays;
