+++ tomcat/src/main/java/org/apache/coyote/http11/file/RequestFactory.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.file;
 
 import java.io.BufferedReader;
 import java.io.IOException;
