+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -2,7 +2,7 @@ package org.apache.coyote.http11.request;
 
 import java.util.Map;
 
-import org.apache.coyote.http11.MimeTypeMaker;
+import org.apache.coyote.http11.file.MimeTypeMaker;
 import org.apache.coyote.http11.request.component.RequestBody;
 import org.apache.coyote.http11.request.component.RequestHeader;
 import org.apache.coyote.http11.request.component.RequestLine;
